+++ toDoMateProject/accounts/views.py
@@ -3,17 +3,17 @@ import datetime
 
 from allauth.socialaccount.models import SocialAccount
 from django.http import JsonResponse, HttpResponseRedirect
-from rest_framework import status, generics
+from rest_framework import status
 from rest_framework.views import APIView
-from rest_framework.permissions import AllowAny, IsAuthenticated
-from .models import User, Task
+from rest_framework.permissions import AllowAny
+from .models import User
 
 # Email Verification
 from allauth.account.models import EmailConfirmation, EmailConfirmationHMAC
 
 # Social Login
 from dj_rest_auth.registration.views import SocialLoginView
-from django.shortcuts import redirect, get_object_or_404
+from django.shortcuts import redirect
 from json import JSONDecodeError
 import os
 import requests
@@ -21,8 +21,6 @@ import requests
 # Google
 from allauth.socialaccount.providers.google.views import GoogleOAuth2Adapter
 from allauth.socialaccount.providers.oauth2.client import OAuth2Client
-
-from .serializers import TaskSerializer, TaskUpdateNameSerializer, TaskUpdateDateSerializer
 
 class ConfirmEmailView(APIView):
     permission_classes = [AllowAny]
@@ -148,75 +146,3 @@ class GoogleToDjangoLogin(SocialLoginView): # if you want to use Authorization C
     adapter_class = GoogleOAuth2Adapter
     callback_url = GOOGLE_CALLBACK_URI
     client_class = OAuth2Client
-
-
-class TaskListCreateView(generics.ListCreateAPIView):
-    def get_queryset(self):
-        uid = self.request.user.id
-        return Task.objects.filter(created_by_id=uid)
-
-    serializer_class = TaskSerializer
-    permission_classes = [IsAuthenticated]
-
-    def perform_create(self, serializer):
-        uid = self.request.user.id
-        serializer.save(created_by_id=uid)
-
-class TaskListView(generics.ListCreateAPIView):
-    def get_queryset(self):
-        uid = self.request.user.id
-        date = self.kwargs['date']
-        return Task.objects.filter(created_by_id=uid, date=date)
-
-    serializer_class = TaskUpdateNameSerializer
-    permission_classes = [IsAuthenticated]
-
-    def perform_create(self, serializer):
-        uid = self.request.user.id
-        date = self.kwargs['date']
-        serializer.save(created_by_id=uid, date=date)
-
-class TaskDetailDestroyView(generics.RetrieveUpdateDestroyAPIView):
-    def get_object(self):
-        uid = self.request.user.id
-        tid = self.kwargs['tid']
-        return get_object_or_404(Task, created_by_id=uid, id=tid)
-
-    serializer_class = TaskSerializer
-    permission_classes = [IsAuthenticated]
-    http_method_names = ['get', 'delete']
-
-class TaskUpdateNameView(generics.RetrieveUpdateDestroyAPIView):
-    def get_object(self):
-        uid = self.request.user.id
-        tid = self.kwargs['tid']
-        return get_object_or_404(Task, created_by_id=uid, id=tid)
-
-    serializer_class = TaskUpdateNameSerializer
-    http_method_names = ['get', 'put']
-
-
-class TaskUpdateDateView(generics.RetrieveUpdateDestroyAPIView):
-    def get_object(self):
-        uid = self.request.user.id
-        tid = self.kwargs['tid']
-        return get_object_or_404(Task, created_by_id=uid, id=tid)
-
-    serializer_class = TaskUpdateDateSerializer
-    http_method_names = ['get', 'put']
-
-def switch_complete(request, *args, **kwargs):
-    uid = request.user.id
-    tid = kwargs.get('tid')
-    task = get_object_or_404(Task, created_by_id=uid, id=tid)
-    task.complete = not task.complete
-    task.save()
-    return redirect(f"http://127.0.0.1:8000/accounts/task/{tid}") #실제 url
-
-def switch_tomorrow(request, *args, **kwargs):
-    uid = request.user.id
-    tid = kwargs.get('tid')
-    task = get_object_or_404(Task, created_by_id=uid, id=tid)
-    task.date = task.date + datetime.timedelta(days=1)
-    task.save()
-    return redirect(f"http://127.0.0.1:8000/accounts/task/{tid}")  # 실제 url
